Add files via upload
In this project, the goal was to predict the stock price of Hascol using various statistical and deep learning models. To start, the data was visualized to get a better understanding of the trend and patterns in the stock price. Data normalization was also performed to ensure the data was on a similar scale.

Five different models were then applied to the data to make predictions, including LSTM Algorithm, Arima Algorithm, HWES (Holt-Winters Exponential Smoothing) Model, Auto Encoder Model, and CNN Model.

After building the models, the performance was evaluated using various evaluation metrics, including Mean Absolute Error (MAE), Mean Squared Error (MSE), Root Mean Squared Error (RMSE), R-squared (R2), Mean Absolute Percentage Error (MAPE), and Symmetric Mean Absolute Percentage Error (SMAPE). The evaluation metrics helped to compare the performance of each of the models and determine which one had the best performance.

The results showed that the performance of the models differed based on the evaluation metric used. By comparing the scores of each evaluation metric, the best-performing model could be identified and further fine-tuned to improve its performance. Overall, the project helped to gain insights into the performance of various models and determine the best model for stock price prediction.